# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🚗 Auto-École API - Authentification Authentik",
        "description": "API complète pour la gestion de l'authentification avec Authentik comme IAM central. Cette API permet l'inscription, la connexion, la gestion des utilisateurs et des tokens avec un système de contournement du Password Grant Flow.",
        "contact": {
            "name": "Support Auto-École API",
            "url": "https://auto-ecole-api.com",
            "email": "support@auto-ecole-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
...
        },
        {
...
        },
...
    "paths": {
...
        "/api/auto-ecoles/{id}/formations": {
            "get": {
                "tags": [
                    "🏫 Auto-Écoles"
                ],
                "summary": "📚 Formations d'une auto-école",
                "description": "Récupère toutes les formations actives proposées par une auto-école",
                "operationId": "getFormationsAutoEcole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Liste des formations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "auto_ecole": {
                                            "type": "object"
                                        },
                                        "formations": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "❌ Auto-école non trouvée"
                    }
                }
            }
        },
...
        "/api/formations": {
            "get": {
                "tags": [
                    "📚 Formations"
                ],
                "summary": "📋 Liste des formations",
                "operationId": "getFormations",
                "parameters": [
                    {
                        "name": "auto_ecole_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Liste des formations"
                    }
                }
            },
            "post": {
                "tags": [
                    "📚 Formations"
                ],
                "summary": "➕ Créer une formation",
                "operationId": "storeFormation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "auto_ecole_id",
                                    "type_permis_id",
                                    "montant",
                                    "session_id"
                                ],
                                "properties": {
                                    "auto_ecole_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "type_permis_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250000
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "session_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "statut": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "✅ Formation créée"
                    },
                    "422": {
                        "description": "❌ Erreur de validation"
                    }
                }
            }
        },
        "/api/formations/{id}": {
            "get": {
                "tags": [
                    "📚 Formations"
                ],
                "summary": "🔍 Détails d'une formation",
                "operationId": "getFormation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Détails de la formation"
                    },
                    "404": {
                        "description": "❌ Formation non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "📚 Formations"
                ],
                "summary": "✏️ Mettre à jour une formation",
                "operationId": "updateFormation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "montant": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "statut": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "✅ Formation mise à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "📚 Formations"
                ],
                "summary": "🗑️ Supprimer une formation",
                "operationId": "deleteFormation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Formation supprimée"
                    }
                }
            }
        },
        "/api/formations/{id}/documents-requis": {
            "get": {
                "tags": [
                    "📚 Formations"
                ],
                "summary": "📋 Documents requis pour une formation",
                "description": "Liste tous les documents obligatoires pour s'inscrire à cette formation",
                "operationId": "getDocumentsRequis",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Liste des documents requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "formation": {
                                            "type": "object"
                                        },
                                        "documents_requis": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "❌ Formation non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Token d'accès JWT obtenu via l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearer_token": {
                "type": "http",
                "description": "Token d'authentification JWT obtenu via /api/auth/login-direct",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}